requirements.txt
openai
pyaudio
speechrecognition
python-dotenv
numpy
sounddevice
scipy